requirements.txt
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
scikit-learn==1.2.2
requests==2.31.0
python-binance==1.0.16
python-dotenv==0.21.0